requirements.txt
langchain-core
langchainhub
langchain-community
langchain-google-genai
langchain-chroma
python-dotenv
langgraph
grandalf
beautifulsoup4
tiktoken